# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "API for Grocery delivery startup",
    "version": "1.0.0",
    "title": "Delivery Service API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "Lists Requests",
      "description": "Delivery Service API for submitting and returning Lists"
    }
  ],
  "paths": {
    "/list/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the user that we want to match",
          "type": "string"
        }
      ],
      "get": {
        "tags": [
          "List Request"
        ],
        "summary": "Get List by user ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "User id",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/itemRequest"
            }
          },
          "400": {
            "description": "Failed. Misunderstood Request."
          },
          "404": {
            "description": "Failed. Item request not found."
          }
        }
      },
      "post": {
        "summary": "create a list by user ID",
        "tags": [
          "List Create Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "User id",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "requestBody": {
          "description": "Item Request Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemRequestBody"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/components/schemas/itemRequest"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      },
      "put": {
        "summary": "Update a list by user ID",
        "tags": [
          "List Update Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "User id",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "requestBody": {
          "description": "Item Request Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemRequestBody"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/itemRequest"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      },
      "delete": {
        "summary": "Clear list by user ID",
        "tags": [
          "List Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Item Request Id",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/itemRequest"
            }
          },
          "404": {
            "description": "Failed. Item Request not found."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "id": {
        "properties": {
          "uuid": {
            "type": "string"
          }
        }
      },
      "list": {
        "properties": {
          "object": {
            "type": "json"
          }
        }
      },
      "ItemRequestBody": {
        "type": "object",
        "properties": {
          "list": {
            "type": "object",
            "format": "json"
          }
        }
      },
      "itemRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "string"
          }
        }
      }
    }
  }
}